Moved utils from ddpg to common utils.py.
