+++ src/rl_agents/common/utils.py
@@ -2,6 +2,8 @@
 """Defines utility functions used across the package."""
 
 import types
+import matplotlib.pyplot as plt
+import numpy as np
 
 
 def inherit_docs(cls):
@@ -17,3 +19,16 @@ def inherit_docs(cls):
                     func.__doc__ = parfunc.__doc__
                     break
     return cls
+
+# @todo: Task for Hassaan to fix this! Sorry Hassaan :D
+def plot_learning(scores, filename, x=None, window=5):
+    n = len(scores)
+    running_avg = np.empty(n)
+    for t in range(n):
+        running_avg[t] = np.mean(scores[max(0, t-window):(t+1)])
+    if x is None:
+        x = [i for i in range(n)]
+    plt.ylabel('Score')
+    plt.xlabel('Game')
+    plt.plot(x, running_avg)
+    plt.savefig(filename)
