--- src/rl_agents/ddpg/utils.py
@@ -1,14 +1,0 @@
-import matplotlib.pyplot as plt 
-import numpy as np
-
-def plot_learning(scores, filename, x=None, window=5):   
-    N = len(scores)
-    running_avg = np.empty(N)
-    for t in range(N):
-	    running_avg[t] = np.mean(scores[max(0, t-window):(t+1)])
-    if x is None:
-        x = [i for i in range(N)]
-    plt.ylabel('Score')       
-    plt.xlabel('Game')                     
-    plt.plot(x, running_avg)
-    plt.savefig(filename)
